extract out the common code from the hierarchical comment definition block builder, which is basically all of it.
